requirements.txt
hydra-core>=1.2.0,<1.3
omegaconf>=2.2,<2.3
pyyaml==6.0.1
torchmetrics>=0.4.1rc0,<=0.10.3
transformers==4.31.0
wandb
webdataset>=0.1.48,<=0.1.62
pandas
sentencepiece
youtokentome
h5py
ijson
matplotlib>=3.3.2
sacremoses
sacrebleu
einops
faiss-cpu
sentence_transformers<=2.2.2
nltk>=3.6.5
jieba
ftfy
gdown
inflect
jieba
opencc
pangu
rapidfuzz
pybind11
GPUtil
pytorch-lightning==1.8.6
ipadic
mecab-python3
protobuf~=3.20
datasets
dill==0.3.7
